Add external classes
These classes may be used by output code.
They are licensed under a different license so that the output code
is not GPLv3.
