+++ Deamer/include/Deamer/External/Cpp/Ast/Node.h
@@ -1,0 +1,15 @@
+#ifndef DEAMER_EXTERNAL_AST_NODE_H
+#define DEAMER_EXTERNAL_AST_NODE_H
+
+namespace deamer::external::cpp::ast
+{
+	class Node
+	{
+	private:
+	public:
+		Node() = default;
+		virtual ~Node() = default;
+	};
+}
+
+#endif // DEAMER_EXTERNAL_AST_NODE_H
